back/requirements.txt
numpy==1.26.4
pandas==2.2.2
requests==2.32.2
Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary==2.9.1
